The config file
All the sensitive information like user name, password, schema, folder name have been redacted. This is the main config file. All the configurations are present in this file. This file is used as driver file. It has to be used in conjunction with static_variables file because this file has the name of all the sections of config files. Config file is created to minimize the hardcoding and to ensure that the environmental variables can be changed without a need to deploy the program again. For more details see https://divijsharma.medium.com/config-files-and-logging-a7b4cc377fd5