Statics Variable file
This file has the name of all static variables. The file is created separately to ensure that all static variables are stored in a single file. Also, in case of any change in the name of section in config file the program with business logic need not be touched at all and only place the relevant change is required is the static variable file, this reduces the human error. This file is used in conjunction with config file. All the sensitive information like folder details have been redacted.